Adds UI fixes for Chng PW and User actions
Adds UI fixes when changing pw.  Now detects if the new pw is the same
as the old and prompts the user to change the new pw.

Adds UI fix for when the user is not an owner of a player resource when
deleting and updating.  UI now tells the user they need to be an
"owner" of the player.